mcp-server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
referencing==0.37.0
rich==14.3.2
rpds-py==0.30.0
shellingham==1.5.4
sse-starlette==3.2.0
starlette==0.52.1
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
